requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.2.7            # via channels, daphne
attrs==19.3.0             # via automat, service-identity, twisted
autobahn==20.4.3          # via daphne
automat==20.2.0           # via twisted
cffi==1.14.0              # via cryptography
channels==2.4.0           # via -r requirements.in
colorlog==4.1.0           # via -r requirements.in
constantly==15.1.0        # via twisted
cryptography==2.9.2       # via autobahn, pyopenssl, service-identity
daphne==2.5.0             # via channels
django-admin-interface==0.12.2  # via -r requirements.in
django-better-admin-arrayfield==1.1.0  # via -r requirements.in
django-colorfield==0.3.0  # via django-admin-interface
django-extensions==2.2.9  # via -r requirements.in
django-flat-responsive==2.0  # via django-admin-interface
django-flat-theme==1.1.4  # via django-admin-interface
django-polymorphic==2.1.2  # via -r requirements.in
django==2.2.12            # via -r requirements.in, channels, django-polymorphic
docutils==0.16            # via -r requirements.in
hyperlink==19.0.0         # via twisted
idna==2.9                 # via hyperlink, twisted
incremental==17.5.0       # via twisted
psycopg2==2.8.5           # via -r requirements.in
pyasn1-modules==0.2.8     # via service-identity
pyasn1==0.4.8             # via pyasn1-modules, service-identity
pycparser==2.20           # via cffi
pyhamcrest==2.0.2         # via twisted
pyopenssl==19.1.0         # via twisted
pytz==2020.1              # via django
service-identity==18.1.0  # via twisted
six==1.14.0               # via automat, cryptography, django-admin-interface, django-extensions, pyopenssl
sqlparse==0.3.1           # via django
twisted[tls]==20.3.0      # via daphne
txaio==20.4.1             # via autobahn
zope.interface==5.1.0     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
